Second hygiene pass: retire Function51BA00 alias, type item cursor global, promote clear bools
Call GiveHeldItemToCharacterOrParty from the portrait path; store g_value_00685072 as
W8ItemInstance*; promote deep_fatigue_applied and several 0/1 predicates to bool.

+++ src/wiz8/local_screens/RCSItemsPage.cpp
@@ -296,7 +296,7 @@ void UseItem005BA4F0(W8ItemInstance* item)
     } else {
         if (IsUsableItemClass00522A00(item) == 0 || IsSpecialItemId004DA0F0(item) != 0) {
             g_flag_00685071 = 1;
-            g_value_00685072 = (int)item;
+            g_value_00685072 = item;
             g_value_00685077 = static_cast<char>(giReviewCharSlot);
             GetOriginOfCharacterItem(giReviewCharSlot, item, &g_flag_00685076, &slot);
         } else {
